server/requirements.txt
Django == 4.0.6
djangorestframework == 3.14.0
numpy == 1.24.0
django-cors-headers
Pillow
django-filter
django-environ
djangorestframework-api-key